Extract more elasticsearch code (#310)
* Move Elasticsearch code under backend package

* Move Elasticsearch code under backend package
